+++ lib/dbus/core_ext/class/attribute.rb
@@ -1,15 +1,17 @@
+# frozen_string_literal: true
 # copied from activesupport/core_ext from Rails, MIT license
-# https://github.com/rails/rails/tree/5aa869861c192daceafe3a3ee50eb93f5a2b7bd2/activesupport/lib/active_support/core_ext
-require 'dbus/core_ext/kernel/singleton_class'
-require 'dbus/core_ext/module/remove_method'
-require 'dbus/core_ext/array/extract_options'
+# https://github.com/rails/rails/tree/9794e85351243cac6d4e78adaba634b8e4ecad0a/activesupport/lib/active_support/core_ext
+
+require "dbus/core_ext/module/redefine_method"
 
 class Class
   # Declare a class-level attribute whose value is inheritable by subclasses.
   # Subclasses can change their own value and it will not impact parent class.
   #
+  # ==== Examples
+  #
   #   class Base
-  #     class_attribute :setting
+  #     my_class_attribute :setting
   #   end
   #
   #   class Subclass < Base
@@ -22,14 +24,14 @@ class Class
   #   Base.setting                # => true
   #
   # In the above case as long as Subclass does not assign a value to setting
-  # by performing <tt>Subclass.setting = _something_ </tt>, <tt>Subclass.setting</tt>
+  # by performing <tt>Subclass.setting = _something_</tt>, <tt>Subclass.setting</tt>
   # would read value assigned to parent class. Once Subclass assigns a value then
   # the value assigned by Subclass would be returned.
   #
   # This matches normal Ruby method inheritance: think of writing an attribute
   # on a subclass as overriding the reader method. However, you need to be aware
   # when using +class_attribute+ with mutable structures as +Array+ or +Hash+.
-  # In such cases, you don't want to do changes in places but use setters:
+  # In such cases, you don't want to do changes in place. Instead use setters:
   #
   #   Base.setting = []
   #   Base.setting                # => []
@@ -59,39 +61,25 @@ class Class
   #   object.setting = false
   #   object.setting          # => false
   #   Base.setting            # => true
-  #
-  # To opt out of the instance reader method, pass <tt>instance_reader: false</tt>.
-  #
-  #   object.setting          # => NoMethodError
-  #   object.setting?         # => NoMethodError
-  #
-  # To opt out of the instance writer method, pass <tt>instance_writer: false</tt>.
-  #
-  #   object.setting = false  # => NoMethodError
-  #
-  # To opt out of both instance methods, pass <tt>instance_accessor: false</tt>.
-  def class_attribute(*attrs)
-    options = attrs.extract_options!
-    instance_reader = options.fetch(:instance_accessor, true) && options.fetch(:instance_reader, true)
-    instance_writer = options.fetch(:instance_accessor, true) && options.fetch(:instance_writer, true)
-    instance_predicate = options.fetch(:instance_predicate, true)
+  def my_class_attribute(*attrs)
+    instance_reader    = true
+    instance_writer    = true
 
     attrs.each do |name|
+      singleton_class.silence_redefinition_of_method(name)
       define_singleton_method(name) { nil }
-      define_singleton_method("#{name}?") { !!public_send(name) } if instance_predicate
 
-      ivar = "@#{name}"
+      ivar = "@#{name}".to_sym
 
+      singleton_class.silence_redefinition_of_method("#{name}=")
       define_singleton_method("#{name}=") do |val|
         singleton_class.class_eval do
-          remove_possible_method(name)
-          define_method(name) { val }
+          redefine_method(name) { val }
         end
 
         if singleton_class?
           class_eval do
-            remove_possible_method(name)
-            define_method(name) do
+            redefine_method(name) do
               if instance_variable_defined? ivar
                 instance_variable_get ivar
               else
@@ -104,26 +92,20 @@ class Class
       end
 
       if instance_reader
-        remove_possible_method name
-        define_method(name) do
+        redefine_method(name) do
           if instance_variable_defined?(ivar)
             instance_variable_get ivar
           else
             self.class.public_send name
           end
         end
-        define_method("#{name}?") { !!public_send(name) } if instance_predicate
       end
 
-      attr_writer name if instance_writer
+      if instance_writer
+        redefine_method("#{name}=") do |val|
+          instance_variable_set ivar, val
+        end
+      end
     end
   end
-
-  private
-
-    unless respond_to?(:singleton_class?)
-      def singleton_class?
-        ancestors.first != self
-      end
-    end
 end