+++ NEWS.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+Bug fixes:
+ * Fixed a conflict with activesupport 5.2 ([#71])
+
+[#71]: https://github.com/mvidner/ruby-dbus/issues/71
+
 ## Ruby D-Bus 0.14.1 - 2018-01-05
 
 Bug fixes: